Revert "fix(proxy): split standalone proxy"
This reverts commit 467b6d4126b57c0459689dc38cf9ce2790f96212.
