+++ packages/config/package.json
@@ -17,7 +17,7 @@
     },
     "homepage": "https://github.com/RedHatInsights/frontend-components/tree/master/packages/config#readme",
     "dependencies": {
-        "@redhat-cloud-services/frontend-components-config-utilities": "1.3.1",
+        "@redhat-cloud-services/frontend-components-config-utilities": "^1.3.0",
         "assert": "^2.0.0",
         "babel-loader": "^8.2.2",
         "browserify-zlib": "^0.2.0",
